Refactor: move common operations to
