+++ dokukratie/helpers/mmmeta.py
@@ -1,11 +1,10 @@
 from datetime import datetime, timedelta
 
 from memorious import settings
+from memorious_extended.util import ensure_date
+from memorious_extended.util import get_env_or_context as _geoc
 from mmmeta import mmmeta
 from servicelayer import env
-
-from .util import ensure_date
-from .util import get_env_or_context as _geoc
 
 
 def get_start_date(context):
