+++ dokukratie/operations/__init__.py
@@ -1,0 +1,2 @@
+from .clean import clean  # noqa
+from .initialize import init  # noqa
